optimization for ios image upload.
