+++ ios/Classes/ImageUploadListener.swift
@@ -1,0 +1,8 @@
+//
+//  ImageUploadListener.swift
+//  amazon_s3_cognito
+//
+//  Created by Paras mac on 11/10/21.
+//
+
+
