+++ pubspec.yaml
@@ -1,5 +1,6 @@
 name: amazon_s3_cognito
-description: This plugin allows users to list files, upload and delete image for amazon s3 cognito
+description: This plugin allows users to upload and delete image for amazon s3 cognito.
+Users can also upload multiple images ti s3 using the plugin.
 version: 0.6.4
 homepage: https://github.com/prachiFam/amazon_s3_cognito
